二次修改

+++ router/index.js
@@ -2,11 +2,26 @@ const router = require('koa-router')();
 const db=require("../db")
 const query=require("../db/query.js")
 
-
+router.post("/api/count",async ctx=>{
+  
+       const aaa= await query("select count(*) from shunli_koa")
+       console.log(aaa.data[0]);
+       ctx.body={
+           code:0,
+           msg:aaa.data[0]
+       }
+     
+})
 router.post("/api/add",async ctx=>{
 
     const {xu,bei,lian,time}=ctx.request.body
-    if(xu&&bei&&lian&&time){
+    const num=await query("select * from shunli_koa where lian=?",[lian])
+    if(num.data.length>=1){
+        ctx.body={
+            code:0,
+            msg:"链接已存在"
+        }
+    }else if(xu&&bei&&lian&&time){
         try{
             const aaa=await query("insert into shunli_koa (xu,bei,lian,time)values(?,?,?,?)",[xu,bei,lian,time])
             ctx.body={
@@ -23,7 +38,7 @@ router.post("/api/add",async ctx=>{
     }else{
         ctx.body={
             code:0,
-            msg:"写全再点"
+            msg:"写全了再点"
         }
     }
 
@@ -31,17 +46,34 @@ router.post("/api/add",async ctx=>{
 
 router.post("/api/delete",async ctx=>{
     const {xu,bei,lian,time,id}=ctx.request.body
-    try{
-        const aaa=await query("delete from shunli_koa where id=?",[id])
-        ctx.body={
-            code:1,
-            msg:"删除成功"
-        }
-    }catch(e){
+    if(id===undefined){
         ctx.body={
             code:0,
-            msg:e.message
+            msg:"传入id"
         }
+    }else{
+        const num=await query("select * from shunli_koa where id=?",[id])
+        if(num.data.length>=1){
+            try{
+                const aaa=await query("delete from shunli_koa where id=?",[id])
+                ctx.body={
+                    code:1,
+                    msg:"删除成功"
+                }
+            }catch(e){
+                ctx.body={
+                    code:0,
+                    msg:e.message
+                }
+            }
+
+        }else{
+            ctx.body={
+                code:1,
+                msg:"蠢货，数据已删除了！你还要删几遍？"
+            }
+        }
+
     }
 
 })
